Fix: Use correct type for alarm ID in NotificationReceiver
Changed the type of the alarm ID from `Long` to `Int` in `NotificationReceiver`
 when sending the `UiEvent.Update` event.

This fixes a potential type mismatch issue and ensures the correct alarm ID is used for updating the alarm state.

+++ services/notifications/src/main/java/city/zouitel/notifications/reciver/NotificationReceiver.kt
@@ -67,6 +67,6 @@ class NotificationReceiver: BroadcastReceiver(), KoinComponent {
             notifyBuilder
         )
 
-        alarmModel.sendUiEvent(UiEvent.Update(intent.getIntExtra(ID, 0).toLong()))
+        alarmModel.sendUiEvent(UiEvent.Update(intent.getIntExtra(ID, 0)))
     }
 }
